+++ src/helpers/tests/sorting-helpers.test.ts
@@ -7,10 +7,11 @@
 import { FOLDERS } from '@zextras/carbonio-ui-commons';
 
 import { getFilterQuery, parseMessageSortingOptions } from 'helpers/sorting';
+import { SortDirection } from '../../types';
 
 const minimalDefaultSortingSettings = {
 	sortType: 'date',
-	sortDirection: 'Desc' as 'Asc' | 'Desc'
+	sortDirection: 'Desc' as SortDirection
 };
 
 test('parseMessageSortingOptions returns the defaultSortingSettings when the sortOrder is not received', () => {
